Reorganize search filters
- Move certificate and yoga styles to secondary filters
- Sort primary filters
- Sort mobile filters in the same order
