+++ src/translations/en.json
@@ -752,7 +752,6 @@
   "SavedCardDetails.replaceCardTitle": "Use another payment method",
   "SavedCardDetails.savedPaymentMethodPlaceholderDesktop": "•••• •••• •••• {last4Digits}",
   "SavedCardDetails.savedPaymentMethodPlaceholderMobile": "…{last4Digits}",
-  "SearchFilters.certificateLabel": "Teacher certificate",
   "SearchFilters.filtersButtonLabel": "Filters",
   "SearchFilters.foundResults": "{count, number} {count, plural, one {result} other {results}}",
   "SearchFilters.keywordLabel": "Keyword",
@@ -762,7 +761,6 @@
   "SearchFilters.noResults": "Could not find any teachers.",
   "SearchFilters.noResultsMobile": "No results.",
   "SearchFilters.openMapView": "Map",
-  "SearchFilters.yogaStylesLabel": "Yoga styles",
   "SearchFiltersMobile.cancel": "CANCEL",
   "SearchFiltersMobile.certificateLabel": "Teacher certificate",
   "SearchFiltersMobile.heading": "Filter teachers",
@@ -773,7 +771,9 @@
   "SearchFiltersPanel.apply": "Apply",
   "SearchFiltersPanel.cancel": "Cancel",
   "SearchFiltersPanel.categoryLabel": "Category",
+  "SearchFiltersPanel.certificateLabel": "Teacher certificate",
   "SearchFiltersPanel.resetAll": "Reset all",
+  "SearchFiltersPanel.yogaStylesLabel": "Yoga styles",
   "SearchMapInfoCard.noImage": "No image",
   "SearchPage.schemaDescription": "Showing search results",
   "SearchPage.schemaMapSearch": "map search",
