+++ src/translations/es.json
@@ -739,7 +739,6 @@
   "SavedCardDetails.replaceCardTitle": "Usar otro método de pago",
   "SavedCardDetails.savedPaymentMethodPlaceholderDesktop": "•••• •••• •••• {last4Digits}",
   "SavedCardDetails.savedPaymentMethodPlaceholderMobile": "…{last4Digits}",
-  "SearchFilters.certificateLabel": "Certificado",
   "SearchFilters.filtersButtonLabel": "Filtros",
   "SearchFilters.foundResults": "{count, number} {count, plural, one {resultado} other {resultados}}",
   "SearchFilters.keywordLabel": "Palabra clave",
@@ -749,7 +748,6 @@
   "SearchFilters.noResults": "No se han encontrado profesores.",
   "SearchFilters.noResultsMobile": "No hay resultados.",
   "SearchFilters.openMapView": "Mapa",
-  "SearchFilters.yogaStylesLabel": "Estilos de yoga",
   "SearchFiltersMobile.cancel": "CANCELAR",
   "SearchFiltersMobile.certificateLabel": "Certificado",
   "SearchFiltersMobile.heading": "Filtrar profesores",
@@ -759,7 +757,9 @@
   "SearchFiltersMobile.yogaStylesLabel": "Estilos de yoga",
   "SearchFiltersPanel.apply": "Aplicar",
   "SearchFiltersPanel.cancel": "Cancelar",
+  "SearchFiltersPanel.certificateLabel": "Certificado",
   "SearchFiltersPanel.resetAll": "Borrar",
+  "SearchFiltersPanel.yogaStylesLabel": "Estilos de yoga",
   "SearchMapInfoCard.noImage": "No hay imagen",
   "SearchPage.schemaDescription": "Mostrando resultados de búsqueda",
   "SearchPage.schemaMapSearch": "búsqueda con mapa",
