+++ src/translations/fr.json
@@ -749,7 +749,6 @@
   "SavedCardDetails.replaceCardTitle": "Utiliser un autre moyen de paiement",
   "SavedCardDetails.savedPaymentMethodPlaceholderDesktop": "•••• •••• •••• {last4Digits}",
   "SavedCardDetails.savedPaymentMethodPlaceholderMobile": "…{last4Digits}",
-  "SearchFilters.certificateLabel": "Diplôme",
   "SearchFilters.filtersButtonLabel": "Filtres",
   "SearchFilters.foundResults": "{count, number} {count, plural, one {résultat} other {résultats}}",
   "SearchFilters.keywordLabel": "Mot-clé",
@@ -759,7 +758,6 @@
   "SearchFilters.noResults": "Aucune professeur trouvé.",
   "SearchFilters.noResultsMobile": "Aucun résultat.",
   "SearchFilters.openMapView": "Carte",
-  "SearchFilters.yogaStylesLabel": "Styles de yoga",
   "SearchFiltersMobile.cancel": "ANNULER",
   "SearchFiltersMobile.certificateLabel": "Diplôme",
   "SearchFiltersMobile.heading": "Filtrer les professeurs",
@@ -769,7 +767,9 @@
   "SearchFiltersMobile.yogaStylesLabel": "Styles de yoga",
   "SearchFiltersPanel.apply": "Valider",
   "SearchFiltersPanel.cancel": "Annuler",
+  "SearchFiltersPanel.certificateLabel": "Diplôme",
   "SearchFiltersPanel.resetAll": "Effacer",
+  "SearchFiltersPanel.yogaStylesLabel": "Styles de yoga",
   "SearchMapInfoCard.noImage": "Aucune image",
   "SearchPage.schemaDescription": "Affichage des résultats de recherche",
   "SearchPage.schemaMapSearch": "recherche sur la carte",
